requirements.txt
flask
bootstrap-flask
flask_sqlalchemy
flask_wtf
wtforms
flask-login
python-dotenv